hoàn thành thiết kế giao diện frmChiTietPhieuXuat Sửa lỗi validate trong frmChiTietPhieuNhap

+++ DAL/ChiTietPhieuNhapDAL.cs
@@ -32,9 +32,9 @@ namespace DAL
         }
         public TrangThai ThemNguyenLieuVaoCTPN(object[] input)
         {
-            if ((int)DatabaseAccess.Instance.ExecuteNonQuery("usp_ThemNguyenLieuVaoCTPN", CachThucHien.StoredProcedure,
-                new string[] { "@MaPhieuNhap", "@MaNguyenLieu", "@DonGia", "@SoLuong", "@HanSuDung" }, 
-                input) 
+            if (DatabaseAccess.Instance.ExecuteNonQuery("usp_ThemNguyenLieuVaoCTPN", CachThucHien.StoredProcedure,
+                new string[] { "@MaPhieuNhap", "@MaNguyenLieu", "@DonGia", "@SoLuong", "@HanSuDung" },
+                input)
                 > 0)
                 return TrangThai.ThanhCong;
             return TrangThai.ThatBai;
@@ -50,11 +50,11 @@ namespace DAL
             return TrangThai.ThatBai;
 
         }
-        public TrangThai XoaNguyenLieuKhoiCTPN(int maPhieuNhap,int maNguyenLieu)
+        public TrangThai XoaNguyenLieuKhoiCTPN(int maPhieuNhap, int maNguyenLieu)
         {
             if ((int)DatabaseAccess.Instance.ExecuteNonQuery("usp_XoaNguyenLieuKhoiCTPN", CachThucHien.StoredProcedure,
                 new string[] { "@MaPhieuNhap", "@MaNguyenLieu" },
-                new object[] {maPhieuNhap,maNguyenLieu})
+                new object[] { maPhieuNhap, maNguyenLieu })
                 > 0)
                 return TrangThai.ThanhCong;
             return TrangThai.ThatBai;
